feat: adding slide buttoms for exit, delete and link
