+++ lib/main.dart
@@ -19,7 +19,8 @@ class MyApp extends StatelessWidget {
         theme: ThemeData.light(),
         darkTheme: ThemeData.dark(),
         themeMode: ThemeMode.system,
-        home: SignIn(),
+        home: Dashboard()
+      //SignIn(),
         );
   }
 }
