+++ lib/purchases/screens/PurchasesLists.dart
@@ -6,6 +6,7 @@ import 'package:finance_controlinator_mobile/components/DefaultInput.dart';
 import '../../authentications/services/AuthorizationService.dart';
 import '../../components/DefaultDialog.dart';
 import '../domain/PurchaseList.dart';
+import 'PurchaseListItem.dart';
 
 class PurchasesListsScreen extends StatelessWidget {
   final listStateKey = GlobalKey<_PurchaseListsState>();
@@ -67,7 +68,6 @@ class PurchasesListsScreen extends StatelessWidget {
 }
 
 class PurchaseLists extends StatefulWidget {
-
   PurchaseLists(Key? key) : super(key: key);
 
   @override
@@ -121,62 +121,5 @@ class _PurchaseListsState extends State<PurchaseLists> {
                       if (purchaseLists == null) return const Text("");
                       return PurchaseListItem(purchaseLists![index]);
                     })));
-  }
-}
-
-class PurchaseListItem extends StatelessWidget {
-  final PurchaseList _purchaseList;
-
-  const PurchaseListItem(PurchaseList purchaseList, {Key? key})
-      : _purchaseList = purchaseList,
-        super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 80,
-      width: double.infinity,
-      decoration: BoxDecoration(
-          color: Colors.lightGreen.shade100,
-          borderRadius: BorderRadius.circular(8),
-          boxShadow: [
-            BoxShadow(
-                color: Colors.grey.withOpacity(.6),
-                spreadRadius: 1,
-                blurRadius: 2,
-                offset: const Offset(2, 2))
-          ]),
-      margin: const EdgeInsets.only(top: 8, bottom: 8, left: 4, right: 4),
-      padding: const EdgeInsets.all(16),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Expanded(
-            flex: 6,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Text(
-                  _purchaseList.name,
-                  overflow: TextOverflow.ellipsis,
-                  maxLines: 2,
-                  style: const TextStyle(
-                      fontWeight: FontWeight.bold, fontSize: 18),
-                )
-              ],
-            ),
-          ),
-          _purchaseList.inProgress
-              ? const Padding(
-                  padding: EdgeInsets.only(left: 20),
-                  child: Icon(
-                    Icons.shopping_cart,
-                    color: Colors.blueAccent,
-                  ))
-              : const SizedBox.shrink(),
-        ],
-      ),
-    );
   }
 }
