+++ pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   intl: ^0.17.0
   flutter_secure_storage: ^5.0.2
   dio: ^4.0.6
+  flutter_slidable: 1.3.0
 
 dev_dependencies:
   flutter_test:
